Simplify logic
#1767
