+++ waltz-ng/client/common/components/boolean-renderer.js
@@ -15,8 +15,8 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import _ from "lodash";
 import {initialiseData} from "../../common";
+import {strToBool} from "../bool-utils";
 
 
 const bindings = {
@@ -24,32 +24,14 @@ const bindings = {
 };
 
 
-const template = `<waltz-icon ng-if="$ctrl.value === true" name="check" class="text-success"></waltz-icon>
-                  <waltz-icon ng-if="$ctrl.value === false" name="times" class="text-danger"></waltz-icon>
-                  <span ng-if="$ctrl.value == null" class="text-muted">-</span>`;
+const template = `<waltz-icon ng-if="$ctrl.booleanValue === true" name="check" class="text-success"></waltz-icon>
+                  <waltz-icon ng-if="$ctrl.booleanValue === false" name="times" class="text-danger"></waltz-icon>
+                  <span ng-if="$ctrl.booleanValue == null" class="text-muted">-</span>`;
 
 
 const initialState = {
     booleanValue: null
 };
-
-
-const aliases = {
-    'TRUE': true,
-    'YES': true,
-    'Y': true,
-    'FALSE': false,
-    'NO': false,
-    'N': false
-};
-
-
-function strToBool(boolStr) {
-    if (boolStr && _.isString(boolStr) && _.has(aliases, _.toUpper(boolStr))) {
-        return aliases[boolStr];
-    }
-    return null;
-}
 
 
 function controller() {
